Move EmptyThreadView into the existing thread route
Instead of a separate /thread/new/page.tsx, the [threadId] route checks
for threadId === 'new' in ThreadPageInner and renders a small inline
EmptyThreadView component (same fixed-header template, "New Thread"
title, no polls, no fetch). The home FAB still navigates to
/thread/new/, which is now served by the same dynamic route.
